Data.Range

Documentation

class (Eq a, Show a, Monoid a, Arbitrary a) => ARanges a where

Minimal complete definition

fromListR, fromListRN, toListR, insertR, deleteR, elemR, insertRange, deleteRange

Associated Types

type RElem a :: *

Methods

fromListR :: [RElem a] -> a

fromListRN :: Int -> [RElem a] -> a

toListR :: a -> [RElem a]

insertR :: RElem a -> a -> a

deleteR :: RElem a -> a -> a

elemR :: RElem a -> a -> Bool

insertRange :: Range (RElem a) -> a -> a

deleteRange :: Range (RElem a) -> a -> a

data Range a

Range of values, equivalent to [a..b] with cases for [], [a..a], and [a..a+b]

Constructors

EmptyRange

Range

Fields

fromRange :: a

Range2

Fields

fromRangeMin :: a

fromRangeMax :: a

minRange :: Range a -> Maybe a

Minimum of a Range

maxRange :: Range a -> Maybe a

Maximum of a Range

foldRange :: Enum a => (a -> b -> b) -> b -> Range a -> b

Left-fold over the Range, using the type's Enum instance to provide values between the endpoints

ppRangeWith :: (a -> String) -> Range a -> String

Pretty-print a Range using the given method to show the endpoints